Replace object with conditional statements
Removed the object that stored the network cardNumber lengths as properties. Instead, I used conditional statements that evaluates length of the cardNumber and then returns the correct network name. If there's a case where there are two networks that have the same number lengths, I then evaluate the first digit of their prefixes.