Modify detectNetwork function, add logic for Discover and Maestro, and debug test file
Modified the detectNetwork function to include a for loop that loops through the cardNumber string and evaluates both the length and the prefix at each loop. If no condition is satisified, we add the current number in the loop to the prefix string.

Included conditional statements to identify Discover and Maestro.

Debugged the test file: modified tests, changed should helper function to expect helper functions