Add variable that creates a string based on cardNumber length
Created a variable 'str' that allows me to add any number of length to a given prefix based on cardNumber. For example, if a card's prefix is '501' and the cardNumber length is 8, str.length would be equal to 5. This gives me the ability to loop through a range of lengths set to a prefix.